requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
pgvector==0.3.6
google-genai==0.3.0

